finance: use unix sockets
We need to start the express app before the checks run.
If we run the check in parallel we do not know which check is
run first and which is run last. Check50 does not provide hooks
for global setup or shutdown handler.

The easiest solution is to an instance of the node app for each check.
To prevent port collision we use unique unix sockets.