finance: check50.flask style methods
Ease the creation of new checks with three new helper methods:
- post: Perform a post request.
- get: Perform a get request.
- status: Test the status code of the last request.
- css_select: Test the html body of the last response contains
              the specified elements.

While there, fix the problem with check dependencies.
A check inherits the filesystem of the check it depends on.
This includes the unix domain socket of the express app.
Remove the socket, so the new instance can create its own socket.