finance: use random users to reset balance
Buying stocks consumes cash. We cannot rely on the sell
functionality.
We need to use a new user for each run to ensure we have enough
balance.

We create a new user in the register check.
To pass the credentials to later checks, we write to the file .env.
It is sourced when the express app starts, and we can get the
credentials from the environment variables.